signup bast

+++ src/main/java/me/jweissen/aeticket/repository/EventRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
     @Query("SELECT e FROM Event e WHERE e.start > CURRENT_TIMESTAMP")
     List<Event> findAllFuture();
 }
